% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{BM010}
\alias{BM010}
\title{Property prices in housing market}
\usage{
BM010(EJKAT20, OMR20, PRIS20, Tid, lang = "en")
}
\arguments{
\item{EJKAT20}{is the property category. Call realestateDK::table_vars("BM010") to see all available parameter settings.}

\item{OMR20}{is the area. Call realestateDK::table_vars("BM010") to see all available parameter settings.}

\item{PRIS20}{is the prices of completed transactions. Call realestateDK::table_vars("BM010") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("BM010") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Property prices in housing market by area, property category, prices of completed transactions and time
}
\examples{
realestateDK::BM010(EJKAT20 = "1", OMR20 = "00", PRIS20 = "UDBUD", Tid = "1992K1")
}
\references{
\url{http://rkr.statistikbank.dk/BM010}
}
