% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{BM011}
\alias{BM011}
\title{Property prices in housing market}
\usage{
BM011(EJKAT20, PNR20, PRIS20, Tid, lang = "en")
}
\arguments{
\item{EJKAT20}{is the property category. Call realestateDK::table_vars("BM011") to see all available parameter settings.}

\item{PNR20}{is the postal code. Call realestateDK::table_vars("BM011") to see all available parameter settings.}

\item{PRIS20}{is the prices of completed transactions. Call realestateDK::table_vars("BM011") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("BM011") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Property prices in housing market by postal code, property category, prices of completed transactions and time
}
\examples{
realestateDK::BM011(EJKAT20 = "1", PNR20 = "1000", PRIS20 = "UDBUD", Tid = "1992K1")
}
\references{
\url{http://rkr.statistikbank.dk/BM011}
}
