% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{BM020}
\alias{BM020}
\title{Movements in housing market}
\usage{
BM020(BEV20, EJKAT20, OMR20, Tid, lang = "en")
}
\arguments{
\item{BEV20}{is the type of movement. Call realestateDK::table_vars("BM020") to see all available parameter settings.}

\item{EJKAT20}{is the property category. Call realestateDK::table_vars("BM020") to see all available parameter settings.}

\item{OMR20}{is the area. Call realestateDK::table_vars("BM020") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("BM020") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Movements in housing market by area, property category, type of movement and time
}
\examples{
realestateDK::BM020(BEV20 = "SALG", EJKAT20 = "1", OMR20 = "00", Tid = "2004K1")
}
\references{
\url{http://rkr.statistikbank.dk/BM020}
}
