% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{BM021}
\alias{BM021}
\title{Movements in housing market}
\usage{
BM021(BEV20, EJKAT20, PNR20, Tid, lang = "en")
}
\arguments{
\item{BEV20}{is the type of movement. Call realestateDK::table_vars("BM021") to see all available parameter settings.}

\item{EJKAT20}{is the property category. Call realestateDK::table_vars("BM021") to see all available parameter settings.}

\item{PNR20}{is the postal code. Call realestateDK::table_vars("BM021") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("BM021") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Movements in housing market by postal code, property category, type of movement and time
}
\examples{
realestateDK::BM021(BEV20 = "SALG", EJKAT20 = "1", PNR20 = "1000", Tid = "2004K1")
}
\references{
\url{http://rkr.statistikbank.dk/BM021}
}
