% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{BM030}
\alias{BM030}
\title{Times-on-market (days)}
\usage{
BM030(EJKAT20, OMR20, Tid, lang = "en")
}
\arguments{
\item{EJKAT20}{is the property category. Call realestateDK::table_vars("BM030") to see all available parameter settings.}

\item{OMR20}{is the area. Call realestateDK::table_vars("BM030") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("BM030") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Times-on-market (days) by area, property category and time
}
\examples{
realestateDK::BM030(EJKAT20 = "1", OMR20 = "00", Tid = "2004K1")
}
\references{
\url{http://rkr.statistikbank.dk/BM030}
}
