% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{BM031}
\alias{BM031}
\title{Times-on-market (days)}
\usage{
BM031(EJKAT20, PNR20, Tid, lang = "en")
}
\arguments{
\item{EJKAT20}{is the property category. Call realestateDK::table_vars("BM031") to see all available parameter settings.}

\item{PNR20}{is the postal code. Call realestateDK::table_vars("BM031") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("BM031") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Times-on-market (days) by postal code, property category and time
}
\examples{
realestateDK::BM031(EJKAT20 = "1", PNR20 = "1000", Tid = "2004K1")
}
\references{
\url{http://rkr.statistikbank.dk/BM031}
}
