% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{table_vars}
\alias{table_vars}
\title{Show all parameter settings}
\usage{
table_vars(table_id)
}
\arguments{
\item{table_id}{the id of the table you want to call.}
}
\value{
a tibble
}
\description{
Show alle the available parameter settings for a given table id.
}
\examples{
realestateDK::table_vars("BM010")

}
