% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{UDB010}
\alias{UDB010}
\title{Housing market}
\usage{
UDB010(BOLA20, EJKAT20, OMR20, Tid, lang = "en")
}
\arguments{
\item{BOLA20}{is the dwellings. Call realestateDK::table_vars("UDB010") to see all available parameter settings.}

\item{EJKAT20}{is the property category. Call realestateDK::table_vars("UDB010") to see all available parameter settings.}

\item{OMR20}{is the area. Call realestateDK::table_vars("UDB010") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("UDB010") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Housing market by area, property category, dwellings and time
}
\examples{
realestateDK::UDB010(BOLA20 = "6", EJKAT20 = "1", OMR20 = "00", Tid = "2004M01")
}
\references{
\url{http://rkr.statistikbank.dk/UDB010}
}
