% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{UDB020}
\alias{UDB020}
\title{Property prices}
\usage{
UDB020(BOLB20, EJKAT20, OMR20, Tid, lang = "en")
}
\arguments{
\item{BOLB20}{is the prices. Call realestateDK::table_vars("UDB020") to see all available parameter settings.}

\item{EJKAT20}{is the property category. Call realestateDK::table_vars("UDB020") to see all available parameter settings.}

\item{OMR20}{is the area. Call realestateDK::table_vars("UDB020") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("UDB020") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Property prices by area, property category, prices and time
}
\examples{
realestateDK::UDB020(BOLB20 = "4", EJKAT20 = "1", OMR20 = "00", Tid = "2004M01")
}
\references{
\url{http://rkr.statistikbank.dk/UDB020}
}
