% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{UDB030}
\alias{UDB030}
\title{Supply and storage times (days)}
\usage{
UDB030(BOLC20, EJKAT20, OMR20, Tid, lang = "en")
}
\arguments{
\item{BOLC20}{is the time on market. Call realestateDK::table_vars("UDB030") to see all available parameter settings.}

\item{EJKAT20}{is the property category. Call realestateDK::table_vars("UDB030") to see all available parameter settings.}

\item{OMR20}{is the area. Call realestateDK::table_vars("UDB030") to see all available parameter settings.}

\item{Tid}{is the time. Call realestateDK::table_vars("UDB030") to see all available parameter settings.}

\item{lang}{whether to return the data in english or danish.}
}
\value{
a tibble
}
\description{
Supply and storage times (days) by area, property category, time on market and time
}
\examples{
realestateDK::UDB030(BOLC20 = "8", EJKAT20 = "1", OMR20 = "00", Tid = "2004M01")
}
\references{
\url{http://rkr.statistikbank.dk/UDB030}
}
